Mppt sim
Simulation of the solar panel battery charging circuit with both c and matlab algorithms working. The calculations of all the components are also calculated and times f times ajusted to resemble real life (fs = 500 Hz, the arduino mega refresh rate, 1e-6 for the buck circuit and 100e-6 for the microcontroller circuit).